Convert Global.spec.js to typescript
